requirement.txt
adbutils==0.16.2
allure-pytest==2.13.1
allure-python-commons==2.13.1
apkutils2==1.0.0
attrs==22.2.0
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cigam==0.0.3
colorama==0.4.6
comtypes==1.1.14
decorator==4.4.2
dependency==0.0.3
Deprecated==1.2.13
deprecation==2.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.9.0
findit==0.5.9
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.1.0
imutils==0.5.4
iniconfig==2.0.0
install==1.3.5
joblib==1.2.0
loguru==0.7.0
logzero==1.7.0
lxml==4.9.2
moviepy==1.0.3
networkx==2.6.3
numpy==1.21.6
opencv-python==4.7.0.72
openpyxl==3.1.1
ordering==0.2.0
packaging==20.9
Pillow==9.4.0
playsound==1.3.0
pluggy==1.0.0
proglog==0.1.10
progress==1.6
psutil==5.9.5
py==1.11.0
PyAudio==0.2.13
pycaw==20230407
pycparser==2.21
pyelftools==0.29
pygame==2.5.2
pyparsing==3.0.9
pypiwin32==223
pyserial==3.5
pytest==7.2.2
pytest-assume==2.4.3
pytest-check==2.1.4
pytest-order==1.1.0
pytest-picked==0.4.6
pytest-random-order==1.1.0
pytest-repeat==0.9.1
pytest-rerunfailures==11.1.2
pyttsx3==2.90
PyWavelets==1.3.0
pywin32==305
PyYAML==6.0
repeat==0.1.3
requests==2.27.1
retry==0.9.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.5.0
uiautomator2==2.16.14
urllib3==1.26.14
wda==0.3.0
websocket==0.2.1
websocket-client==0.59.0
whichcraft==0.6.1
win32-setctime==1.1.0
wrapt==1.15.0
XlsxWriter==3.0.8
xmltodict==0.13.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
